SocketInfoを socket_info.hpp に移動
